change field to not auto

+++ server/accountapp/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class Group(models.Model):
     name = models.CharField(max_length=30, blank=False)
     admin = models.ForeignKey(User, on_delete=models.CASCADE)
-    hash = models.BigAutoField()
+    hash = models.IntegerField()
 
     def __str__(self):
         return str(self.name)
